// src/components/Pagination/Pagination.types.ts
import type { HTMLAttributes } from "react";

export type PaginationVariant =
    | "solid"
    | "outline"
    | "ghost";

export type PaginationSize =
    | "sm"
    | "md"
    | "lg";

export type PaginationRounded =
    | "none"
    | "sm"
    | "md"
    | "full";

export interface PaginationProps
    extends HTMLAttributes<HTMLDivElement> {
    page: number;
    totalPages: number;
    onPageChange: (page: number) => void;
    siblingCount?: number;
    showFirstLast?: boolean;
    showPrevNext?: boolean;
    variant?: PaginationVariant;
    size?: PaginationSize;
    rounded?: PaginationRounded;
    fullWidth?: boolean;
    disabled?: boolean;
}